+++ src/main/java/com/group/libraryapp/domain/book/BookRepository.java
@@ -1,5 +1,6 @@
 package com.group.libraryapp.domain.book;
 
+import com.group.libraryapp.domain.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
