+++ src/main/java/com/group/libraryapp/domain/book/JavaBook.java
@@ -8,7 +8,7 @@ import javax.persistence.Id;
 import static javax.persistence.GenerationType.IDENTITY;
 
 @Entity
-public class Book {
+public class JavaBook {
 
   @Id
   @GeneratedValue(strategy = IDENTITY)
@@ -17,11 +17,11 @@ public class Book {
   @Column(nullable = false)
   private String name;
 
-  public Book() {
+  public JavaBook() {
 
   }
 
-  public Book(String name) {
+  public JavaBook(String name) {
     if (name.isBlank()) {
       throw new IllegalArgumentException("이름은 비어 있을 수 없습니다");
     }
