+++ src/main/java/com/group/libraryapp/domain/user/User.java
@@ -1,6 +1,6 @@
 package com.group.libraryapp.domain.user;
 
-import com.group.libraryapp.domain.book.Book;
+import com.group.libraryapp.domain.Book;
 import com.group.libraryapp.domain.user.loanhistory.UserLoanHistory;
 
 import javax.persistence.*;
