+++ src/main/java/com/group/libraryapp/service/book/BookService.java
@@ -1,6 +1,6 @@
 package com.group.libraryapp.service.book;
 
-import com.group.libraryapp.domain.book.Book;
+import com.group.libraryapp.domain.Book;
 import com.group.libraryapp.domain.book.BookRepository;
 import com.group.libraryapp.domain.user.User;
 import com.group.libraryapp.domain.user.UserRepository;
@@ -30,7 +30,7 @@ public class BookService {
 
   @Transactional
   public void saveBook(BookRequest request) {
-    Book newBook = new Book(request.getName());
+    Book newBook = new Book(request.getName(), null);
     bookRepository.save(newBook);
   }
