+++ src/test/kotlin/com/group/libraryapp/service/BookServiceTest.kt
@@ -1,6 +1,6 @@
 package com.group.libraryapp.service
 
-import com.group.libraryapp.domain.book.Book
+import com.group.libraryapp.domain.Book
 import com.group.libraryapp.domain.book.BookRepository
 import com.group.libraryapp.domain.user.User
 import com.group.libraryapp.domain.user.UserRepository
